+++ CHANGELOG.md
@@ -7,6 +7,12 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+### Changed
+
+- The mobile tab devices now default to opening closed. This way the user can
+  determine when to first open them or open them externally. This makes the
+  behavior closer to what the instances tab does.
+
 ## [0.13.0][] - 2022-06-30
 
 ### Added
@@ -277,7 +283,5 @@ release date when you use `npm version` (see `README.md`).
 [0.12.0]: https://github.com/webxdc/webxdc-dev/tree/v0.12.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.12.1...HEAD
 [0.12.1]: https://github.com/webxdc/webxdc-dev/tree/v0.12.1
-
-
-[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.13.0...HEAD
+[unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.13.0...HEAD
 [0.13.0]: https://github.com/webxdc/webxdc-dev/tree/v0.13.0